🎨: domain 패키지를 사용하여 프로젝트 구조 재구성
- 핵심 비지니스 로직을 담당하는 클래스를 domain 패키지에 추가하여 구성하였습니다.
