+++ README.md
@@ -21,36 +21,39 @@
   ```bash
   📘 class 📗 interface
   
-  ├── 📁src
-  │   ├── 📁main
-  │   │    ├── 📁application
-  │   │    │   ├── 📘Application.java
-  │   │    ├── 📁controller
-  │   │    │   ├── 📘LadderController.java
-  │   │    ├── 📁message
-  │   │    │   ├── 📘GameMessage.java  
-  │   │    │   └── 📘ShapeMessage.java
-  │   │    ├── 📁model.game
-  │   │    │   ├── 📁ladder
-  │   │    │   │   ├── 📁element
-  │   │    │   │   │   ├── 📗Element.java
-  │   │    │   │   │   ├── 📘Line.java
-  │   │    │   │   │   ├── 📘None.java
-  │   │    │   │   │   └── 📘Pole.java
-  │   │    │   │   ├── 📘Ladder.java
-  │   │    │   │   └── 📘Shape.java
-  │   │    │   ├── 📁player
-  │   │    │   │   └── 📘Player.java
-  │   │    │   └── 📘Game.java
-  │   │    ├── 📁util
-  │   │    │   ├── 📘GameUtils.java 
-  │   │    │   ├── 📘Parser.java 
-  │   │    │   ├── 📘RandomUtils.java 
-  │   │    │   └── 📘StringUtils.java
-  │   │    ├── 📁view
-  │   │    │   ├── 📘InputView.java
-  │   │    │   └── 📘OutputView.java
-  └── └── 📘Main.java
+  ├──📁 main
+  │  ├──📁 java
+  │  │  ├──📁 application
+  │  │  │  ├──📁 controller
+  │  │  │  │  └──📘 LadderController.java
+  │  │  │  ├──📁 domain
+  │  │  │  │  ├──📁ladder
+  │  │  │  │  │     ├──📁element
+  │  │  │  │  │     │  ├──📗Element.java
+  │  │  │  │  │     │  ├──📘Line.java
+  │  │  │  │  │     │  ├──📘None.java
+  │  │  │  │  │     │  └──📘Pole.java
+  │  │  │  │  │     ├──📘Ladder.java
+  │  │  │  │  │     └──📘Shape.java
+  │  │  │  │  └──📁player
+  │  │  │  │     └──📘Player.java
+  │  │  │  ├──📁 game
+  │  │  │  │  └──📘Game.java
+  │  │  │  ├──📁message
+  │  │  │  │  ├──📘GameMessage.java  
+  │  │  │  │  └──📘ShapeMessage.java
+  │  │  │  ├──📁 util
+  │  │  │  │  ├──📘 GameUtils.java 
+  │  │  │  │  ├──📘 Parser.java 
+  │  │  │  │  ├──📘 RandomUtils.java 
+  │  │  │  │  └──📘 StringUtils.java
+  │  │  │  ├──📁validator
+  │  │  │  │  └──📘 InputValidator.java 
+  │  │  │  ├──📁 view
+  │  │  │  │  ├──📘 InputView.java
+  │  │  │  │  └──📘 OutputView.java
+  │  │  │  └──📘 Application.java
+  └──└──└──📘 Main.java
   ```
   
 - ### 프로젝트 설계 및 주요 코드 설명
