+++ src/main/java/application/domain/ladder/element/Element.java
@@ -1,0 +1,5 @@
+package application.domain.ladder.element;
+
+public interface Element {
+
+}
