+++ src/main/java/application/domain/ladder/element/Pole.java
@@ -1,0 +1,10 @@
+package application.domain.ladder.element;
+
+public class Pole implements Element {
+    private final String str = "|";
+
+    @Override
+    public String toString() {
+        return str;
+    }
+}
