+++ src/main/java/application/domain/player/Player.java
@@ -1,0 +1,14 @@
+package application.domain.player;
+
+public class Player {
+    private final String name;
+
+    public Player(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String toString() {
+        return name;
+    }
+}
