+++ src/main/java/application/util/RandomUtils.java
@@ -1,0 +1,11 @@
+package application.util;
+
+import java.util.Random;
+
+public class RandomUtils {
+    private static Random random = new Random();
+
+    public static double nextDouble() {
+        return random.nextDouble();
+    }
+}
