+++ src/main/java/application/util/StringUtils.java
@@ -1,0 +1,7 @@
+package application.util;
+
+public class StringUtils {
+    public static String rpad(int pad, String player) {
+        return String.format("%-" + pad + "s", player);
+    }
+}
