+++ src/main/java/application/validator/InputValidator.java
@@ -1,0 +1,11 @@
+package application.validator;
+
+import java.util.Arrays;
+
+public class InputValidator {
+    public static final int MAX_NAME_LENGTH = 5;
+
+    public static boolean validateNameLength(String[] names) {
+        return Arrays.stream(names).anyMatch((name) -> name.length() > MAX_NAME_LENGTH);
+    }
+}
