--- src/main/java/message/GameMessage.java
@@ -1,8 +1,0 @@
-package message;
-
-public class GameMessage {
-    public static final String REQUEST_PLAYER_NAMES = "참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분하세요)";
-    public static final String REQUEST_LADDER_HEIGHT = "최대 사다리 높이는 몇 개인가요?";
-    public static final String PLAYER_NAME_LENGTH_ERROR_MESSAGE = "플레이어 명은 5글자를 초과할 수 없습니다.";
-    public static final String RUNTIME_ERROR_MESSAGE = "게임 실행 중 에러가 발생했습니다.";
-}
