--- src/main/java/model/game/ladder/element/Element.java
@@ -1,5 +1,0 @@
-package model.game.ladder.element;
-
-public interface Element {
-
-}
