--- src/main/java/model/game/ladder/element/Pole.java
@@ -1,10 +1,0 @@
-package model.game.ladder.element;
-
-public class Pole implements Element {
-    private final String str = "|";
-
-    @Override
-    public String toString() {
-        return str;
-    }
-}
