--- src/main/java/model/game/player/Player.java
@@ -1,14 +1,0 @@
-package model.game.player;
-
-public class Player {
-    private final String name;
-
-    public Player(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return name;
-    }
-}
