--- src/main/java/util/RandomUtils.java
@@ -1,11 +1,0 @@
-package util;
-
-import java.util.Random;
-
-public class RandomUtils {
-    private static Random random = new Random();
-
-    public static double nextDouble() {
-        return random.nextDouble();
-    }
-}
