--- src/main/java/util/StringUtils.java
@@ -1,7 +1,0 @@
-package util;
-
-public class StringUtils {
-    public static String rpad(int pad, String player) {
-        return String.format("%-" + pad + "s", player);
-    }
-}
