--- src/main/java/view/OutputView.java
@@ -1,24 +1,0 @@
-package view;
-
-import model.game.Game;
-
-public class OutputView {
-    private static OutputView ov;
-
-    private OutputView() {}
-
-    public static OutputView getInstance() {
-        if (ov == null) {
-            ov = new OutputView();
-        }
-        return ov;
-    }
-
-    public void printGame(Game game) {
-        System.out.println(game);
-    }
-
-    public void printErrMsg(String msg) {
-        System.out.println(msg);
-    }
-}
